+++ src/Eccube/Service/SchemaService.php
@@ -36,10 +36,14 @@ class SchemaService
         $this->entityManager = $entityManager;
     }
 
-    public function updateSchema($generatedFiles, $proxiesDirectory)
+    public function executeCallback(callable $callback, $generatedFiles, $proxiesDirectory, $outputDir = null)
     {
-        $outputDir = sys_get_temp_dir().'/proxy_'.StringUtil::random(12);
-        mkdir($outputDir);
+        $createOutputDir = false;
+        if (is_null($outputDir)) {
+            $outputDir = sys_get_temp_dir().'/proxy_'.StringUtil::random(12);
+            mkdir($outputDir);
+            $createOutputDir = true;
+        }
 
         try {
             $chain = $this->entityManager->getConfiguration()->getMetadataDriverImpl();
@@ -61,13 +65,30 @@ class SchemaService
 
             $tool = new SchemaTool($this->entityManager);
             $metaData = $this->entityManager->getMetadataFactory()->getAllMetadata();
-            $tool->updateSchema($metaData, true);
+
+            call_user_func($callback, $tool, $metaData);
         } finally {
-            foreach (glob("${outputDir}/*") as  $f) {
-                unlink($f);
+            if ($createOutputDir) {
+                foreach (glob("${outputDir}/*") as $f) {
+                    unlink($f);
+                }
+                rmdir($outputDir);
             }
-            rmdir($outputDir);
         }
+    }
+
+    public function updateSchema($generatedFiles, $proxiesDirectory, $saveMode = false)
+    {
+        $this->executeCallback(function (SchemaTool $tool, $metaData) use ($saveMode) {
+            $tool->updateSchema($metaData, $saveMode);
+        }, $generatedFiles, $proxiesDirectory);
+    }
+
+    public function createSchema($generatedFiles, $proxiesDirectory)
+    {
+        $this->executeCallback(function (SchemaTool $tool, $metaData) {
+            $tool->createSchema($metaData);
+        }, $generatedFiles, $proxiesDirectory);
     }
 
     /**
